+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## [0.1.0] - 2024-04-03
 
-- Initial release
+- Initial release, placeholder gem
+
+## [0.2.0] - tbd
+- adds `ChatCompletion` module
+- adds `PromptDeclarations` module
+- adds `FunctionDispatch` module